Add a blocks file that gathers data about block to reuse in our tests
