+++ tests/e2e/specs/backend/attribute-filter.test.js
@@ -12,20 +12,22 @@ import {
 	selectBlockByName,
 } from '@woocommerce/blocks-test-utils';
 
-const block = {
-	name: 'Filter Products by Attribute',
-	slug: 'woocommerce/attribute-filter',
-	class: '.wc-block-attribute-filter',
-};
+/**
+ * Internal dependencies
+ */
+import { filterProductsByAttribute as block } from '../../blocks';
 
 describe( `${ block.name } Block`, () => {
 	beforeAll( async () => {
 		await switchUserToAdmin();
 		await visitBlockPage( `${ block.name } Block` );
 
-		await expect( page ).toClick( 'span.woocommerce-search-list__item-name', { text: 'Capacity' } );
+		await expect( page ).toClick(
+			'span.woocommerce-search-list__item-name',
+			{ text: 'Capacity' }
+		);
 		//needed for attributes list to load correctly
-		await page.waitFor(1000);
+		await page.waitFor( 1000 );
 
 		await expect( page ).toClick( 'button', { text: 'Done' } );
 		await page.waitForNetworkIdle();
@@ -118,7 +120,6 @@ describe( `${ block.name } Block`, () => {
 				// Capacity has only 2 attributes
 			).toEqual( 2 );
 
-
 			await expect( page ).toClick(
 				'span.woocommerce-search-list__item-name',
 				{
@@ -126,7 +127,7 @@ describe( `${ block.name } Block`, () => {
 				}
 			);
 			//needed for attributes list to load correctly
-			await page.waitFor(1000);
+			await page.waitFor( 1000 );
 
 			// reset
 			await expect( page ).toClick(
@@ -136,7 +137,7 @@ describe( `${ block.name } Block`, () => {
 				}
 			);
 			//needed for attributes list to load correctly
-			await page.waitFor(1000);
+			await page.waitFor( 1000 );
 		} );
 
 		it( 'renders on the frontend', async () => {
