+++ tests/e2e/specs/backend/products-by-attribute.test.js
@@ -2,16 +2,13 @@
  * External dependencies
  */
 import { getAllBlocks, switchUserToAdmin } from '@wordpress/e2e-test-utils';
-
 import { visitBlockPage } from '@woocommerce/blocks-test-utils';
 
+/**
+ * Internal dependencies
+ */
 import { insertBlockDontWaitForInsertClose } from '../../utils.js';
-
-const block = {
-	name: 'Products by Attribute',
-	slug: 'woocommerce/products-by-attribute',
-	class: '.wc-block-products-by-attribute',
-};
+import { productsByAttribute as block } from '../../blocks';
 
 describe( `${ block.name } Block`, () => {
 	beforeAll( async () => {
