+++ tests/e2e/specs/backend/single-product.test.js
@@ -3,17 +3,16 @@
  */
 import { getAllBlocks, switchUserToAdmin } from '@wordpress/e2e-test-utils';
 import { insertBlockDontWaitForInsertClose } from '../../utils.js';
+
+/**
+ * Internal dependencies
+ */
 import { visitBlockPage } from '@woocommerce/blocks-test-utils';
+import { singleProduct as block } from '../../blocks';
 
 if ( process.env.WOOCOMMERCE_BLOCKS_PHASE < 3 )
 	// eslint-disable-next-line jest/no-focused-tests
 	test.only( 'skipping all other things', () => {} );
-
-const block = {
-	name: 'Single Product',
-	slug: 'woocommerce/single-product',
-	class: '.wc-block-single-product',
-};
 
 describe( `${ block.name } Block`, () => {
 	beforeAll( async () => {
